requirements.txt
appnope==0.1.4
asttokens==3.0.0
certifi==2025.4.26
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dotenv==0.9.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fonttools==4.57.0
geopandas==1.0.1
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
neo4j==5.28.1
nest-asyncio==1.6.0
numpy<2
#numpy==2.2.5
openpyxl==3.1.5
#packaging==25.0
packaging>=20,<25
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow>=7.1.0,<11
#pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
shapely==2.1.0
six==1.17.0
stack-data==0.6.3
streamlit-folium==0.25.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
